+++ src/sync.rs
@@ -7,9 +7,10 @@ mod prefetch;
 mod progress;
 
 use console::Emoji;
-use std::path::PathBuf;
+use std::collections::BTreeMap;
+use std::path::{Path, PathBuf};
 
-use crate::config::schema::Config;
+use crate::config::schema::{Config, ConfigAsset};
 use crate::config::toml;
 
 use self::install::Installer;
@@ -17,15 +18,24 @@ use self::prefetch::prefetch;
 use self::progress::SyncProgress;
 use self::progress::ToolPair;
 
-pub fn sync_from_path(config_path: PathBuf) {
-    toml::with_parsed_file(config_path, sync_from_config)
+pub fn sync_from_path(config_path: PathBuf, tool: Option<String>) {
+    toml::with_parsed_file(config_path.clone(), |config| {
+        sync_from_config(config, config_path, tool)
+    });
 }
 
-pub fn sync_from_config(config: Config) {
+pub fn sync_from_config(mut config: Config, config_path: PathBuf, tool: Option<String>) {
     if config.tools.is_empty() {
-        no_tools_message()
-    } else {
-        sync_from_config_no_check(config)
+        no_tools_message();
+        return;
+    }
+
+    match tool {
+        Some(tool) => match config.tools.remove(&tool) {
+            Some(asset) => sync_single_tool(config, tool, asset),
+            None => tool_not_in_config_message(&tool, &config_path),
+        },
+        None => sync_from_config_no_check(config),
     }
 }
 
@@ -53,8 +63,24 @@ For more details, refer to the official documentation:
     );
 }
 
+fn tool_not_in_config_message(tool: &str, path: &Path) {
+    eprintln!(
+        r#"The '{}' tool is not listed in the configuration file: {}
+
+Add the tool to the configuration file or use the 'tool install' command for 
+installing one of the tools natively supported by 'tool-sync'."#,
+        tool,
+        path.display(),
+    );
+}
+
 const DONE: Emoji<'_, '_> = Emoji("✨ ", "* ");
 const DIRECTORY: Emoji<'_, '_> = Emoji("📁 ", "* ");
+
+pub fn sync_single_tool(mut config: Config, name: String, asset: ConfigAsset) {
+    config.tools = BTreeMap::from([(name, asset)]);
+    sync_from_config_no_check(config);
+}
 
 /// Like `sync_from_config` but expects non-empty list of tools
 pub fn sync_from_config_no_check(config: Config) {